Self-check for single-instance mode
Instances of the app can now check for other running instances and either proceed or halt based on their status. The presence of a "healthy" instance (i.e. a running instance plus ffplay/omxplayer) causes the new instance to terminate, while the presence of an "unhealthy" instance causes the new instance to terminate rogue processes before proceeding. The lack of any running processes causes execution to proceed nomally.

All of this is controlled by a new app setting called "ForceSingleInstance", which defaults to false.

This behavior is intended to provide clients with a cross-platform way to safely and simply keep the application alive. This can be used in concert with a simple cron job.